requirements.txt
absl-py==0.7.1
alabaster==0.7.12
alembic==1.0.10
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
bokeh==1.1.0
boto==2.49.0
boto3==1.9.169
botocore==1.12.169
catboost==0.14.2
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
colorcet==2.0.1
contractions==0.0.18
cryptography==2.7
cycler==0.10.0
dask==1.2.2
datashader==0.7.0
datashape==0.5.2
decorator==4.4.0
defusedxml==0.6.0
distributed==1.28.1
docutils==0.14
entrypoints==0.3
enum34==1.1.6
filelock==3.0.12
ftfy==5.5.1
gast==0.2.2
gensim==3.7.3
graphviz==0.10.1
grpcio==1.21.1
h5py==2.9.0
HeapDict==1.0.0
holoviews==1.12.3
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.17
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
jedi==0.13.3
jeepney==0.4
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-book==0.5.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kaggle==1.5.4
Keras==2.2.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras-vggface==0.5
keyring==19.0.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lightgbm==2.2.3
llvmlite==0.28.0
locket==0.2.0
Mako==1.0.11
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
more-itertools==7.0.0
msgpack==0.6.1
multipledispatch==0.6.0
nbclean==0.3.2
nbconvert==5.5.0
nbformat==4.4.0
nbgrader==0.5.5
networkx==2.3
notebook==5.7.8
numba==0.43.1
numpy==1.16.3
numpydoc==0.9.1
opencv-python==4.1.0.25
packaging==19.0
pandas==0.24.2
pandoc==1.0.2
pandocfilters==1.4.2
param==1.9.0
parso==0.4.0
partd==0.3.10
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pkg-resources==0.0.0
pluggy==0.12.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.2
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyct==0.4.6
pydot==1.4.1
pyemd==0.5.1
pyflakes==2.1.1
Pygments==2.4.1
pylint==2.3.1
pypandoc==1.4
pyparsing==2.4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
PyQtWebEngine==5.12.1
pyrsistent==0.15.2
pytest==4.6.0
python-dateutil==2.8.0
python-editor==1.0.4
python-slugify==3.0.2
pytz==2019.1
pyviz-comms==0.7.2
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
QtAwesome==0.5.7
qtconsole==4.5.0
QtPy==1.8.0
requests==2.22.0
rope==0.14.0
ruamel.yaml==0.15.96
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
smart-open==1.8.4
snowballstemmer==1.2.1
sortedcontainers==2.1.0
Sphinx==2.1.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.4
spyder-kernels==0.4.4
spyder-notebook==0.1.4
SQLAlchemy==1.3.4
tblib==1.4.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.3.1
text-unidecode==1.2
toml==0.10.0
toolz==0.9.0
torch==1.1.0
tornado==6.0.2
tox==3.12.1
tqdm==4.32.1
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.3
xarray==0.12.1
xgboost==0.90
zict==0.1.4
zipp==0.5.1
